% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{aggregate_adcp}
\alias{aggregate_adcp}
\title{Aggregate columns by by the interval specified by user}
\usage{
aggregate_adcp(x, by = c("month", "year")[1])
}
\arguments{
\item{x}{tibble of data}

\item{by}{char the interval over which to aggregate}
}
\value{
aggregated tibble
}
\description{
Aggregate columns by by the interval specified by user
}
