% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{aggregate_buoy}
\alias{aggregate_buoy}
\title{Aggregate columns by interval}
\usage{
aggregate_buoy(x, by = c("month", "year")[1])
}
\arguments{
\item{x}{tibble of data}

\item{by}{char the interval over which to aggregate}
}
\value{
summary tibble
}
\description{
Aggregate columns by interval
}
