% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{aggregate_optics}
\alias{aggregate_optics}
\title{Aggregate optics by the user specifed interval}
\usage{
aggregate_optics(x, by = c("month", "year")[1])
}
\arguments{
\item{x}{tibble of data}

\item{by}{cahr the interval over which to aggregate}
}
\value{
aggregated pixel
}
\description{
Aggregate optics by the user specifed interval
}
