% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsc.R
\name{aggregate_rtsc}
\alias{aggregate_rtsc}
\title{Aggregate columns by by the interval specified by user}
\usage{
aggregate_rtsc(x, by = c("year", "month")[2])
}
\arguments{
\item{x}{tibble of data}

\item{by}{char the interval over which to aggregate}
}
\value{
aggregated tibble
}
\description{
Aggregate columns by by the interval specified by user
}
