% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{buoy_lut}
\alias{buoy_lut}
\title{Returns a table of buoy metadata}
\usage{
buoy_lut(
  filename = system.file("buoy/buoy_listing.csv", package = "nerddap"),
  form = c("table", "sf")[1]
)
}
\arguments{
\item{filename}{the name of the file to read}

\item{form}{char, one of 'table' or 'sf'}
}
\value{
tibble
}
\description{
Returns a table of buoy metadata
}
