% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{complete_intervals_buoys}
\alias{complete_intervals_buoys}
\title{Clip a table so that only complete intervals are present (for subsequent
aggregation).  We count unique days in the interval, even though data may
ne recorded much more frequently}
\usage{
complete_intervals_buoys(
  x,
  by = c("year", "month")[1],
  min_count = c(year = 364, month = 28)[[by]]
)
}
\arguments{
\item{x}{tibble of buoy data}

\item{by}{character, one of "year" (default) or "month"}

\item{min_count}{numeric defaults to 364 or 365, but adjust to 28 or 29 for month}
}
\value{
tibble clipped to include only complete intervals
}
\description{
Clip a table so that only complete intervals are present (for subsequent
aggregation).  We count unique days in the interval, even though data may
ne recorded much more frequently
}
