% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctd_read_raw}
\alias{ctd_read_raw}
\title{Read raw ctd data for a given buoy}
\usage{
ctd_read_raw(filename)
}
\arguments{
\item{filename}{char, the path to the data to read}
}
\value{
tibble of data
}
\description{
Read raw ctd data for a given buoy
}
