% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{date_to_week}
\alias{date_to_week}
\title{Convert to as POSIXct to an 8D week number (1-46) or 7D week number (1-52)}
\usage{
date_to_week(x, week_length = 7)
}
\arguments{
\item{x}{POSIXct or Date vector}

\item{week_length}{numeric, the number of days per week}
}
\value{
char 01-46 or 01-52 week number
}
\description{
Convert to as POSIXct to an 8D week number (1-46) or 7D week number (1-52)
}
