% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{decompose_direction}
\alias{decompose_direction}
\title{Decompose a direction to u,v vectors}
\usage{
decompose_direction(x, varname = "wind_direction")
}
\arguments{
\item{x}{tibble of data}

\item{varname}{char the name of the varoable to decompose}
}
\value{
tibble with add varname_u and varname_v columns added
}
\description{
Decompose a direction to u,v vectors
}
