% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{drop_suffix}
\alias{drop_suffix}
\title{Given a raw data table, drop 'name_qc' columns}
\usage{
drop_suffix(x, suffix = "_qc")
}
\arguments{
\item{x}{tibble of raw data}

\item{suffix}{character the pattern to identify columns to be dropped}
}
\value{
the input table with selected columns dropped
}
\description{
Given a raw data table, drop 'name_qc' columns
}
