% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{fetch_buoy_adcp}
\alias{fetch_buoy_adcp}
\title{Fetch met data for a given buoy}
\usage{
fetch_buoy_adcp(buoy = "B01")
}
\arguments{
\item{buoy}{char, the buoy id}
}
\value{
tibble of data
}
\description{
Fetch met data for a given buoy
}
