% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{fetch_buoy_ctd}
\alias{fetch_buoy_ctd}
\title{Fetch met data for a given buoy - saves monthly and yearly aggregates}
\usage{
fetch_buoy_ctd(buoy = "B01")
}
\arguments{
\item{buoy}{char, the buoy id}
}
\value{
tibble of data
}
\description{
Fetch met data for a given buoy - saves monthly and yearly aggregates
}
