% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{fetch_buoy_optics}
\alias{fetch_buoy_optics}
\title{Fetch met data for a given buoy (aggregates by year and month)}
\usage{
fetch_buoy_optics(buoy = "B01")
}
\arguments{
\item{buoy}{char, the buoy id}
}
\value{
tibble of data (monthly)
}
\description{
Fetch met data for a given buoy (aggregates by year and month)
}
