% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{fetch_buoys}
\alias{fetch_buoys}
\title{Fetch buoy data for one or more datasets}
\usage{
fetch_buoys(
  buoy = buoy_lut()$id,
  what = c("met", "ctd", "rtsc", "optics", "adcp")
)
}
\arguments{
\item{buoy}{char, one or more buoy id (B01, etc)}

\item{what}{char, one or more datasets names}
}
\description{
Fetch buoy data for one or more datasets
}
