% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{mask_qc}
\alias{mask_qc}
\title{Given a raw data table, for each 'name_qc' mask with a relacement
wherever qc fails threshold}
\usage{
mask_qc(x, threshold = 0, suffix = "_qc", replacement = NA_real_)
}
\arguments{
\item{x}{tibble of raw data}

\item{threshold}{numeric records at or below this value are retained}

\item{suffix}{character the pattern to identify columns to be dropped}

\item{replacement}{numeric or NA, the value to substitute in where qc is bad}
}
\value{
the input table masked
}
\description{
Given a raw data table, for each 'name_qc' mask with a relacement
wherever qc fails threshold
}
