% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{met_aggregate}
\alias{met_aggregate}
\title{Aggregate columns by interval}
\usage{
met_aggregate(x, by = c("month", "year")[1])
}
\arguments{
\item{x}{tibble of data}

\item{by}{char the interval overwich to aggregate}
}
\value{
summary tibble
}
\description{
Aggregate columns by interval
}
