% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{met_read_raw}
\alias{met_read_raw}
\title{Read raw met data for a given buoy}
\usage{
met_read_raw(filename)
}
\arguments{
\item{filename}{char, the path to the data to read}
}
\value{
tibble of data
}
\description{
Read raw met data for a given buoy
}
