% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerddap-package.R
\docType{package}
\name{nerddap}
\alias{nerddap}
\alias{nerddap-package}
\title{nerddap package}
\description{
R-language tools for working with NERACOOS ERDDAP data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/nerddap}
}

}
