% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcp.R
\name{read_raw_adcp}
\alias{read_raw_adcp}
\title{Read raw adcp data for a given buoy}
\usage{
read_raw_adcp(filename)
}
\arguments{
\item{filename}{char, the path to the data to read}
}
\value{
tibble of data
}
\description{
Read raw adcp data for a given buoy
}
