% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{read_raw_optics}
\alias{read_raw_optics}
\title{Read raw optics data for a given buoy}
\usage{
read_raw_optics(filename)
}
\arguments{
\item{filename}{char, the path to the data to read}
}
\value{
tibble of data
}
\description{
Read raw optics data for a given buoy
}
