% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsc.R
\name{read_raw_rtsc}
\alias{read_raw_rtsc}
\title{Read raw rtsc data for a given buoy}
\usage{
read_raw_rtsc(filename)
}
\arguments{
\item{filename}{char, the path to the data to read}
}
\value{
tibble of data
}
\description{
Read raw rtsc data for a given buoy
}
