% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sixnum}
\alias{sixnum}
\title{Given a set of values compute the standard summary of
\code{min, q25, median, mean, q75, max}}
\usage{
sixnum(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector of values}

\item{na.rm}{logical, if TRUE remove NA values before computing summaries}
}
\value{
named vector of summary values
}
\description{
Given a set of values compute the standard summary of
\code{min, q25, median, mean, q75, max}
}
