% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{table_range}
\alias{table_range}
\title{Compute the range of values per column}
\usage{
table_range(x, na.rm = TRUE, collapse = TRUE)
}
\arguments{
\item{x}{a table with one or more numeric columns}

\item{na.rm}{logical, if TRUE compute after removing NAs}

\item{collapse}{logical, if TRUE compute a single range for all data}
}
\value{
if \code{collapse} is TRUE then a vector of min and max
otherwise three column summary of \code{var, min, max}
}
\description{
Compute the range of values per column
}
