% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{time_to_utc}
\alias{time_to_utc}
\title{Modify a timezone attribute on a POSIX object}
\usage{
time_to_utc(x = Sys.time())
}
\arguments{
\item{x}{POSIX time (default current time)}

\item{input}{with timezone attribute modified}
}
\description{
Modify a timezone attribute on a POSIX object
}
